Cosigner cluster work
Cosigner nodes now need to use the same cluster key to communicate with each other. These are random private keys.
Nodes will now propagate their known hosts, if a node can't be reached with the beacon then it can be added manually and it will discover the rest of the network.
Inter-node communication is not secure yet, still in plain text.