Ethereum and FIAT modules redesigned.
ethereum becomes ethereum-core, primarily dealing the connection to the
blockchain and its api.

fiat becomes ethereum-token, which provides a multi-sig contract with
internal addressing. It should be able to work with any standard token as
long as they use the transfer and approve/transferFrom standards.

ethereum-token also takes over as the ethereum multi-sig contract,
treating it like a token with slight modifications to the default token
storage contract.

Contract creation on ethereum is now monitored for receipt. Existing contract
lookup is not 100% reliable at the moment and shouldn't be relied on anyway. It's
recommended that you make note of the contract addreses as Cosigner
creates them (or you do so manually), then insert them into the configuration.

Transaction scanning has been reworked in ethereum modules, Cosigner now
makes use of the logs/filter functionality in ethereum to track these more
efficiently.